requirements.txt
Django==5.0.4
mysqlclient==2.2.4
django-environ==0.11.2
django-cas-ng==5.0.0
openpyxl==3.1.2
reportlab==4.1.0
Pillow==10.3.0
whitenoise==6.6.0
gunicorn==22.0.0
